[Storage-file] Move high level functions to FileClient
