+++ sdk/storage/storage-file/README.md
@@ -51,14 +51,14 @@ There are differences between Node.js and browsers runtime. When getting start w
   - `generateAccountSASQueryParameters()`
   - `generateFileSASQueryParameters()`
 - Parallel uploading and downloading
-  - `uploadFileToAzureFile()`
-  - `uploadStreamToAzureFile()`
-  - `downloadAzureFileToBuffer()`
+  - `FileClient.uploadFile()`
+  - `FileClient.uploadStream()`
+  - `FileClient.downloadToBuffer()`
 
 ##### Following features, interfaces, classes or functions are only available in browsers
 
 - Parallel uploading and downloading
-  - `uploadBrowserDataToAzureFile()`
+  - `FileClient.uploadBrowserData()`
 
 ## Getting Started
 
@@ -124,12 +124,8 @@ The Azure Storage SDK for JavaScript provides low-level and high-level APIs.
 
 ```javascript
 const {
-  Aborter,
   StorageClient,
   FileServiceClient,
-  ShareClient,
-  DirectoryClient,
-  FileClient,
   SharedKeyCredential,
   AnonymousCredential
 } = require("@azure/storage-file");
