+++ sdk/storage/storage-file/samples/javascript/basic.js
@@ -7,7 +7,7 @@ const {
   FileServiceClient,
   SharedKeyCredential,
   AnonymousCredential
-} = require(".."); // Change to "@azure/storage-file" in your package
+} = require("../.."); // Change to "@azure/storage-file" in your package
 
 async function main() {
   // Enter your storage account name and shared key
