+++ sdk/storage/storage-file/samples/typescript/advanced.ts
@@ -4,11 +4,8 @@
 
 import fs from "fs";
 import {
-  AnonymousCredential,
-  downloadAzureFileToBuffer,
-  uploadFileToAzureFile,
-  uploadStreamToAzureFile,
   Aborter,
+  AnonymousCredential,
   FileServiceClient,
   StorageClient
 } from "../.."; // Change to "@azure/storage-file" in your package
@@ -48,21 +45,20 @@ async function main() {
   const fileClient = directoryClient.createFileClient(fileName);
   const fileSize = fs.statSync(localFilePath).size;
 
-  // Parallel uploading with uploadFileToAzureFile in Node.js runtime
-  // uploadFileToAzureFile is only available in Node.js
-  await uploadFileToAzureFile(localFilePath, fileClient, {
+  // Parallel uploading with FileClient.uploadFile() in Node.js runtime
+  // FileClient.uploadFile() is only available in Node.js
+  await fileClient.uploadFile(localFilePath, {
     rangeSize: 4 * 1024 * 1024, // 4MB range size
     parallelism: 20, // 20 concurrency
     progress: (ev) => console.log(ev)
   });
-  console.log("uploadFileToAzureFile success");
+  console.log("uploadFile success");
 
-  // Parallel uploading a Readable stream with uploadStreamToAzureFile in Node.js runtime
-  // uploadStreamToAzureFile is only available in Node.js
-  await uploadStreamToAzureFile(
+  // Parallel uploading a Readable stream with FileClient.uploadStream() in Node.js runtime
+  // FileClient.uploadStream() is only available in Node.js
+  await fileClient.uploadStream(
     fs.createReadStream(localFilePath),
     fileSize,
-    fileClient,
     4 * 1024 * 1024,
     20,
     {
@@ -70,13 +66,13 @@ async function main() {
       progress: (ev: any) => console.log(ev)
     }
   );
-  console.log("uploadStreamToAzureFile success");
+  console.log("uploadStream success");
 
-  // Parallel uploading a browser File/Blob/ArrayBuffer in browsers with uploadBrowserDataToAzureFile
-  // Uncomment following code in browsers because uploadBrowserDataToAzureFile is only available in browsers
+  // Parallel uploading a browser File/Blob/ArrayBuffer in browsers with FileClient.uploadBrowserData()
+  // Uncomment following code in browsers because FileClient.uploadBrowserData() is only available in browsers
   /*
   const browserFile = document.getElementById("fileinput").files[0];
-  await uploadBrowserDataToAzureFile(browserFile, fileClient, {
+  await fileClient.uploadBrowserData(browserFile, {
     rangeSize: 4 * 1024 * 1024, // 4MB range size
     parallelism: 20, // 20 concurrency
     progress: ev => console.log(ev)
@@ -84,15 +80,15 @@ async function main() {
   */
 
   // Parallel downloading an Azure file into Node.js buffer
-  // downloadAzureFileToBuffer is only available in Node.js
+  // FileClient.downloadToBuffer() is only available in Node.js
   const buffer = Buffer.alloc(fileSize);
-  await downloadAzureFileToBuffer(buffer, fileClient, 0, undefined, {
+  await fileClient.downloadToBuffer(buffer, 0, undefined, {
     abortSignal: Aborter.timeout(30 * 60 * 1000),
     rangeSize: 4 * 1024 * 1024, // 4MB range size
     parallelism: 20, // 20 concurrency
     progress: (ev) => console.log(ev)
   });
-  console.log("downloadAzureFileToBuffer success");
+  console.log("downloadToBuffer success");
 
   // Delete share
   await shareClient.delete();
