+++ sdk/storage/storage-file/src/index.ts
@@ -19,9 +19,7 @@ export * from "./ShareClient";
 export * from "./credentials/AnonymousCredential";
 export * from "./credentials/Credential";
 export * from "./credentials/SharedKeyCredential";
-export * from "./highlevel.browser";
 export * from "./highlevel.common";
-export * from "./highlevel.node";
 export { IPRange as IPRange } from "./IPRange";
 export { Range } from "./Range";
 export * from "./Pipeline";
